+++ src/chart/series.rs
@@ -1,9 +1,58 @@
 use super::ChartContext;
 use crate::coord::CoordTranslate;
-use crate::drawing::backend::{BackendCoord, DrawingErrorKind};
-use crate::drawing::{DrawingAreaErrorKind, DrawingBackend};
-use crate::element::{EmptyElement, IntoDynElement, MultiLineText, Rectangle};
+use crate::drawing::DrawingAreaErrorKind;
+use crate::element::{DynElement, EmptyElement, IntoDynElement, MultiLineText, Rectangle};
 use crate::style::{IntoFont, IntoTextStyle, ShapeStyle, SizeDesc, TextStyle, TRANSPARENT};
+
+use plotters_backend::{BackendCoord, DrawingBackend, DrawingErrorKind};
+
+type SeriesAnnoDrawFn<'a, DB> = dyn Fn(BackendCoord) -> DynElement<'a, DB, BackendCoord> + 'a;
+
+/// The annotations (such as the label of the series, the legend element, etc)
+/// When a series is drawn onto a drawing area, an series annotation object
+/// is created and a mutable reference is returned.
+pub struct SeriesAnno<'a, DB: DrawingBackend> {
+    label: Option<String>,
+    draw_func: Option<Box<SeriesAnnoDrawFn<'a, DB>>>,
+}
+
+impl<'a, DB: DrawingBackend> SeriesAnno<'a, DB> {
+    #[allow(clippy::option_as_ref_deref)]
+    pub(crate) fn get_label(&self) -> &str {
+        // TODO: Change this when we bump the MSRV
+        self.label.as_ref().map(|x| x.as_str()).unwrap_or("")
+    }
+
+    pub(crate) fn get_draw_func(&self) -> Option<&SeriesAnnoDrawFn<'a, DB>> {
+        self.draw_func.as_ref().map(|x| x.as_ref())
+    }
+
+    pub(crate) fn new() -> Self {
+        Self {
+            label: None,
+            draw_func: None,
+        }
+    }
+
+    /// Set the series label
+    /// - `label`: The string would be use as label for current series
+    pub fn label<L: Into<String>>(&mut self, label: L) -> &mut Self {
+        self.label = Some(label.into());
+        self
+    }
+
+    /// Set the legend element creator function
+    /// - `func`: The function use to create the element
+    /// *Note*: The creation function uses a shifted pixel-based coordinate system. And place the
+    /// point (0,0) to the mid-right point of the shape
+    pub fn legend<E: IntoDynElement<'a, DB, BackendCoord>, T: Fn(BackendCoord) -> E + 'a>(
+        &mut self,
+        func: T,
+    ) -> &mut Self {
+        self.draw_func = Some(Box::new(move |p| func(p).into_dyn()));
+        self
+    }
+}
 
 /// Describes where we want to put the series label
 pub enum SeriesLabelPosition {
@@ -148,9 +197,9 @@ impl<'a, 'b, DB: DrawingBackend + 'a, CT: CoordTranslate> SeriesLabelStyle<'a, '
             label_element.push_line(label_text);
         }
 
-        let (mut w, mut h) = label_element
-            .estimate_dimension()
-            .map_err(|e| DrawingAreaErrorKind::BackendError(DrawingErrorKind::FontError(e)))?;
+        let (mut w, mut h) = label_element.estimate_dimension().map_err(|e| {
+            DrawingAreaErrorKind::BackendError(DrawingErrorKind::FontError(Box::new(e)))
+        })?;
 
         let margin = self.margin as i32;
 
@@ -178,7 +227,9 @@ impl<'a, 'b, DB: DrawingBackend + 'a, CT: CoordTranslate> SeriesLabelStyle<'a, '
 
         for (((_, y0), (_, y1)), make_elem) in label_element
             .compute_line_layout()
-            .map_err(|e| DrawingAreaErrorKind::BackendError(DrawingErrorKind::FontError(e)))?
+            .map_err(|e| {
+                DrawingAreaErrorKind::BackendError(DrawingErrorKind::FontError(Box::new(e)))
+            })?
             .into_iter()
             .zip(funcs.into_iter())
         {